Fix inspection issues
